+++ libs/application/templates/p-sign/src/forms/applicationWithActor.ts
@@ -1,0 +1,24 @@
+import { buildForm } from '@island.is/application/core'
+import { Form, FormModes } from '@island.is/application/types'
+import { sectionDataProviders } from './applicationSections/sectionDataProviders'
+import { sectionInformationActor } from './applicationSections/sectionInformationActor'
+import { sectionPhoto } from './applicationSections/sectionPhoto'
+import { sectionDelivery } from './applicationSections/sectionDelivery'
+import { sectionOverview } from './applicationSections/sectionOverview'
+
+export const getApplication = (): Form => {
+  return buildForm({
+    id: 'PMarkApplicationDraftForm',
+    title: '',
+    mode: FormModes.APPLYING,
+    renderLastScreenButton: true,
+    renderLastScreenBackButton: true,
+    children: [
+      sectionDataProviders,
+      sectionInformationActor,
+      sectionPhoto,
+      sectionDelivery,
+      sectionOverview,
+    ],
+  })
+}
