+++ libs/application/templates/p-sign/src/lib/constants.ts
@@ -8,6 +8,7 @@ export enum States {
 }
 export enum Roles {
   APPLICANT = 'applicant',
+  ACTOR = 'actor',
 }
 
 export const YES = 'yes'
