+++ libs/application/templates/p-sign/src/lib/pSignTemplate.ts
@@ -35,7 +35,7 @@ const PSignTemplate: ApplicationTemplate<
           },
           progress: 0.33,
           lifecycle: {
-            shouldBeListed: false,
+            shouldBeListed: true,
             shouldBePruned: true,
             whenToPrune: 24 * 3600 * 1000,
           },
@@ -49,6 +49,21 @@ const PSignTemplate: ApplicationTemplate<
               id: Roles.APPLICANT,
               formLoader: () =>
                 import('../forms/application').then((val) =>
+                  Promise.resolve(val.getApplication()),
+                ),
+              actions: [
+                {
+                  event: DefaultEvents.SUBMIT,
+                  name: 'Staðfesta',
+                  type: 'primary',
+                },
+              ],
+              write: 'all',
+            },
+            {
+              id: Roles.ACTOR,
+              formLoader: () =>
+                import('../forms/applicationWithActor').then((val) =>
                   Promise.resolve(val.getApplication()),
                 ),
               actions: [
@@ -78,6 +93,11 @@ const PSignTemplate: ApplicationTemplate<
               formLoader: () => import('../forms/done').then((val) => val.done),
               read: 'all',
             },
+            {
+              id: Roles.ACTOR,
+              formLoader: () => import('../forms/done').then((val) => val.done),
+              read: 'all',
+            },
           ],
         },
         type: 'final' as const,
@@ -88,7 +108,12 @@ const PSignTemplate: ApplicationTemplate<
     nationalId: string,
     application: Application,
   ): ApplicationRole | undefined {
-    if (application.applicant === nationalId) {
+    if (
+      application.applicant === nationalId &&
+      application.applicantActors.length > 0
+    ) {
+      return Roles.ACTOR
+    } else if (application.applicant === nationalId) {
       return Roles.APPLICANT
     }
   },
